TG-54: implement index on Share folder and refactor routes of components
