+++ src/Components/Employees/index.js
@@ -2,8 +2,7 @@ import { useEffect } from 'react';
 import { useSelector, useDispatch } from 'react-redux';
 import { getEmployees } from 'redux/employees/thunks';
 import styles from './employees.module.css';
-import Table from 'Components/Share/Table';
-import Spinner from 'Components/Share/Spinner';
+import { Table, Spinner } from 'Components/Share';
 
 function Employees() {
   const { list, isFetching, error } = useSelector((state) => state.employees);
