+++ src/Components/Share/index.js
@@ -1,0 +1,9 @@
+import Table from 'Components/Share/Table';
+import Form from 'Components/Share/Form';
+import Modal from 'Components/Share/Modal';
+import Spinner from 'Components/Share/Spinner';
+import InputForm from 'Components/Share/InputForm';
+import SelectForm from 'Components/Share/SelectForm';
+import TeamMembersTable from 'Components/Share/TeamMembersTable';
+
+export { Table, Form, Modal, Spinner, InputForm, SelectForm, TeamMembersTable };
