+++ src/Components/SuperAdmins/index.js
@@ -2,8 +2,7 @@ import { useEffect } from 'react';
 import { useSelector, useDispatch } from 'react-redux';
 import { getSuperAdmins } from 'redux/super-admins/thunks';
 import styles from './super-admins.module.css';
-import Table from 'Components/Share/Table';
-import Spinner from 'Components/Share/Spinner';
+import { Table, Spinner } from 'Components/Share';
 
 function SuperAdmins() {
   const { list, isFetching, error } = useSelector((state) => state.superAdmins);
