+++ src/Components/Tasks/index.js
@@ -2,8 +2,7 @@ import { useEffect } from 'react';
 import { useSelector, useDispatch } from 'react-redux';
 import { getTasks } from 'redux/tasks/thunks';
 import styles from './tasks.module.css';
-import Table from 'Components/Share/Table';
-import Spinner from 'Components/Share/Spinner';
+import { Table, Spinner } from 'Components/Share';
 
 function Tasks() {
   const { list, isFetching, error } = useSelector((state) => state.tasks);
