+++ src/routes/projects.jsx
@@ -1,6 +1,6 @@
 import { BrowserRouter as Router, Switch, Route } from 'react-router-dom';
 import Projects from 'Components/Projects';
-import Form from 'Components/Share/Form';
+import { Form } from 'Components/Share';
 
 const ProjectsRoutes = () => {
   return (
