+++ src/routes/tasks.jsx
@@ -1,6 +1,6 @@
 import { BrowserRouter as Router, Switch, Route } from 'react-router-dom';
 import Tasks from 'Components/Tasks';
-import Form from 'Components/Share/Form';
+import { Form } from 'Components/Share';
 
 const TasksRoutes = () => {
   return (
